+++ src/main/kotlin/com/zepben/evolve/database/sqlite/tables/iec61970/base/wires/generation/production/TablePowerElectronicsWindUnit.kt
@@ -7,9 +7,8 @@
  */
 package com.zepben.evolve.database.sqlite.tables.iec61970.base.wires.generation.production
 
-@Suppress("PropertyName")
 class TablePowerElectronicsWindUnit : TablePowerElectronicsUnit() {
-    
+
     override fun name(): String {
         return "power_electronics_wind_unit"
     }