+++ src/test/kotlin/com/zepben/evolve/cim/iec61970/base/wires/ProtectedSwitchTest.kt
@@ -53,5 +53,5 @@ internal class ProtectedSwitchTest {
             ProtectedSwitch::clearOperatedByProtectionEquipment
         )
     }
-    
+
 }